% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is_integer.R
\name{is_integer}
\alias{is_integer}
\title{Determine if number is integer}
\usage{
is_integer(x, val = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
describe output of function
}
\description{
This function determines if x is an integer.
}
\examples{
#description of examples here
#actual R code to execute examples
}
\author{
Ross Whippo
}

