% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qzipf.R
\name{qzipf}
\alias{qzipf}
\title{Quantile of the Zipf distribution}
\usage{
qzipf(p, N, s, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{N}{the number of elements.}

\item{s}{the exponent characterizing the distribution}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{logical;}{if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}
}
\value{
describe output of function
}
\description{
This function generates quantiles from the Zipf distribution via the VGAM pzipf() function.
}
\examples{
#description of examples here
#actual R code to execute examples
}
\author{
piklprado
}

