% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sad_beta.R
\name{sad_beta}
\alias{sad_beta}
\title{short description of what function does}
\usage{
sad_beta(spp.count, Nsamp, Nplot = 16, Nsite = 9, dist = "qlnorm",
  method = "bray", gamma.step = 0, radplot = TRUE, betaplot = TRUE,
  csvoutput = TRUE)
}
\arguments{
\item{spp.count}{number of species to use in simulation}

\item{Nsamp}{number of individuals to sample in each plot}

\item{Nplot}{number of 'plots' sampled per round}

\item{Nsite}{number of 'sites' to sample, essentially number of simulations to run}

\item{dist}{distribution to use for calculating RAD}

\item{method}{beta measurement used}

\item{gamma.step}{integer for stepwise increase of gamma with each progressive site sampled}

\item{radplot}{Logical. If TRUE RAD plot will be generated}

\item{betaplot}{Logical. If TRUE beta plot will be generated}

\item{csvoutput}{Logical. If TRUE the entire data set is exported into .csv in working directory}
}
\value{
describe output of function
}
\description{
much longer description of what function does
}
\examples{
#description of examples here
#actual R code to execute examples
}
\author{
Ross Whippo
}

