% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sad_rad.R
\name{sad_rad}
\alias{sad_rad}
\title{RAD Null Models}
\usage{
sad_rad(spp_count, Nsamp, Nplot = 16, Nsite = 9, dist = "qlnorm")
}
\arguments{
\item{spp_count}{number of species to use in simulation}

\item{Nsamp}{number of individuals to sample in each plot}

\item{Nplot}{number of 'plots' sampled per round}

\item{Nsite}{number of 'sites' to sample, essentially number of simulations to run}

\item{dist}{distribution to use for calculating RAD}

\item{gamma_step}{integer for stepwise increase of gamma with each progressive site sampled}
}
\value{
Uses linear model to fit line to rank abundance distribution, plots and provides slope.
}
\description{
This function randomly populates community matrices from various species abundance distributions (SADs),
based on input parameters chosen by the user.
}
\examples{
#description of examples here
#actual R code to execute examples
}
\author{
Ross Whippo
}

