\name{saDist-package}
\alias{saDist-package}
\alias{saDist}
\docType{package}
\title{
Species Area Distribution Null Models and Visualization
}
\description{
The saDist package was created to do really awesome things.  
}
\details{
\tabular{ll}{
Package: \tab saDist\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2014-12-17\cr
License: \tab MIT\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Ross Whippo

Maintainer: Ross Whippo <rosspitality@gmail.com>
}
\references{
Kraft et al. 2011
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
