% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{saDist}
\alias{saDist}
\alias{saDist-package}
\title{Use species area distributions to create null models.}
\description{
saDist allows you to parameterize null models from different species abundance distributions.
It also provides anlaysis options.
}
\details{
The only function available for this version of saDist is the sad_rad function.
}

