% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sad_beta.R
\name{sad_beta}
\alias{sad_beta}
\title{Beta Diversity Null Models}
\usage{
sad_beta(spp_count, Nsamp, Nplot = 16, Nsite = 9, dist = "lnorm",
  method = "bray")
}
\arguments{
\item{spp_count}{number of species to use in simulation}

\item{Nsamp}{number of individuals to sample in each plot}

\item{Nplot}{number of 'plots' sampled per round}

\item{Nsite}{number of 'sites' to sample, essentially number of simulations to run}

\item{dist}{distribution to use for calculating RAD}

\item{method}{beta measurement used}
}
\value{
Output includes beta diversity plots of all permuted communities.
}
\description{
This function randomly populates community matrices from various species abundance distributions (SADs),
based on input parameters chosen by the user for beta diversity comparison.
}
\examples{
#description of examples here
#actual R code to execute examples
}
\author{
Ross Whippo
}

